Create sandboxed filesystem with bind mounts and symlinks
In the process of doing this, we added a new dependency on the
`libmount` crate. Although the lovely `BindMount` and `Remount` types
exposed by this library were unusable within the user namespace (they
implicitly try to access `/proc`, which does not exist in our
namespace), the included `/proc/self/mountinfo` parser was incredibly
useful.

Up until this point, I found it hard to reconcile my more cross-platform
notion of `enable_sysctl` with the Linux-specific concepts of unsharing
and access to `/proc` and `/dev`. With this new commit, the definition
has become clearer: this flag will permit access to `/proc` and `/dev`.
If this flag is turned off, then binaries will lose the ability to
access `/proc` and `/dev` and PID namespacing will also be enabled.

Since the concept of `sysctl` is so general and poorly defined, it might
be beneficial to split this into a few different flags in the future.
Unfortunately, Bastille plans to support macOS via Seatbelt/sandboxd in
the future, and that operating system has neither `procfs` and PID
namespacing. It's not clear how to expose these semantic differences in
a cross-platform API, but it will be worked out as time progresses.