Add allow_dev_read() method separate from sysctl
Previously, the API mistakenly conflated the definition of device files
(hailing from the realm of `ioctl()` and `/dev`) with `sysctl` (which
actually relates to `sysctl()`, `/proc`, PID namespaces, and kernel
configuration). The API avoided distinguishing between the two for fear
of sacrificing cross-platform compatibility, but in fact, the boundary
between `ioctl()` and `sysctl()` are actually pretty well defined across
essentially all Unix-likes (except for macOS, which does away with the
physical `/proc` filesystem and instead only exposes the `sysctl()`
system call).

Because of this distinction, the builder API is updated to explicitly
allow or deny device access separately from the `sysctl` setting.