Disallow bind mounting procfs if allow_sysctl is false
This is possible by inspecting the `mountinfo` table and looking at the
`fstype` field for the given mounted device. Combined with the changes
from the previous commit, it appears that the `allow_sysctl` and
`allow_dev_read` settings finally have some teeth (although
`allow_sysctl` already controls PID namespacing, but this additional
change strengthens the security on Linux).