Record platform-specific caveats in PLATFORM.md
After having conduncted a survey of sandboxing and kernel capabilities
available in multiple Unix-like operating systems, it should be
seriously evaluated whether we should consider Darwin/macOS as a
first-class citizen. The sandboxing and filesystem capabilities on that
platform are quite crippled, and it would be almost impossible to
maintain feature parity with that platform specifically. We could
potentially reduce the scope of Bastille to the lowest common
denominator, but that seems like a shame when both Linux and FreeBSD
already support the nicest goodies.