Implement PoC chroot on macOS with sandboxfs
This commit adds a dependency on [sandboxfs] for a reasonably performant
equivalent bind mounts that exists on macOS, compared to copying and
hard linking files into a temporary directory every time a new sandboxed
process is started. This unfortunately requires users to install
[osxfuse] before using this library, simply because macOS does not have
any cheap equivalent to bind mounts on Linux or nullfs mounts on
FreeBSD.

[sandboxfs]: https://github.com/bazelbuild/sandboxfs
[osxfuse]: https://osxfuse.github.io/

In this design, the parent process creates a new temporary directory
with an `mnt` subdirectory to use as a sandboxfs mount point. Placed
alongside it is a `sandboxfs.fifo` UNIX named pipe, used for IPC between
this process and the long-running `sandboxfs::mount()` function. The
parent process forks once to create a child process which waits for the
filesystem mounts to become available, and once they appear, the child
chroots itself into that `mnt` directory, enters the macOS sandbox, and
calls `Command::exec()`. The parent process then forks again to create a
long-running child process which sets up sandboxfs and waits for the
first child process to terminate, at which point it down sandboxfs and
cleans up the per-process temporary directory after itself.

A cleaner way to implement this in the future could be to create a
`OnceCell` which contains a hypothetical `SandboxFs` struct with the
cleanup code in its `Drop` implementation. When `os::create_sandbox()`
is called, this `SANDBOXFS` cell is called with
`OnceCell::get_or_init()` and sandboxfs is spawned in a separate thread
and initialized in some temporary directory with no filesystem mappings
specified by default. It would use two pairs of unnamed UNIX pipes for
communication instead of a named pipe to avoid leaving physical traces
on the hard drive. Rather than starting and stopping sandboxfs over and
over again in a new temporary directory every time
`os::create_sandbox()` is called, we would keep sandboxfs running
continuously in the background after the first invocation and
reconfigure the filesystem layout on each subsequent call.