Set UID/GID in macOS sandbox, if specified
This commit implements support for the `uid()` and `gid()` methods in
the macOS backend for Bastille.

According to the sandboxfs man page, the mounted filesystem is only
accessible by the uid that mounted it, which is problematic for us
because we require the root access while setting up the sandbox, and
dropping the sandboxed process down to an unprivileged user later on
would render the chrooted filesystem totally unreadable by that process.

To work around this limitation, we make two crucial changes. First, we
temporarily drop the parent process down to the user-specified UID and
GID, if any, before calling `tempfile::tempdir()` to ensure it is
created with the correct user and group, not owned by `root`. Next, we
enable the `allow_root` FUSE option, which allows the mounted filesystem
to be readable by both the root user (during `chroot` time) and also by
the non-root user assigned to the process within the sandbox.