Use UnixStream to signal to child that sandboxfs is mounted
This commit replaces the static `std::thread::sleep()` delay with an
`std::os::unix::net::UnixStream` for signaling the child process that
sandboxfs is mounted and ready for `chroot` to be called. Here we opt to
use `UnixStream` over `ipc-channel` because it behaves more predictably
than the latter on macOS and is less complex to use.