Fix blocking in sandboxfs subprocess with custom UID/GID
It turns out that `kill(2)` occasionally returns `EPERM` for the
sandboxed child process when its UID/GID is being modified with
`setuid()` and `setgid()`, which would break the sandboxfs mounting
process. This commit accounts for the occasional `EPERM` in the macOS
backend by ignoring the error and not aborting the mounting process when
it occurs.