Use pthread_setugid_np() in "sandboxfs" thread on macOS
There exists a FreeBSD and macOS specific POSIX extension named
`pthread_setugid_np()` which behaves identically to `setuid()` and
`setgid()` except it applies per-thread rather than per-process. This
commit makes use of this extension to ensure that only the "sandboxfs"
thread escalates its UID to 0 instead of the entire subprocess.

Not only does this make the forked sandboxfs subprocess a bit more
secure, it also opens the possibility of eliminating the `fork()` call
entirely and perhaps launching `sandboxfs` in a thread rather than in a
subprocess. This would make Bastille more [async-friendly] on macOS,
bringing it more in line with the Linux backend.

[async-friendly]: https://github.com/tokio-rs/tokio/issues/1541