Fix mounting root dir behavior on macOS
Currently, it is impossible to mount `/` from the host to `/` in the
sandbox on macOS using sandboxfs because `/` can only be configured on
startup via `sandboxfs::mount()` and cannot be reconfigured at runtime
via the input/output pipes. This behavior is possible on Linux.

This commit changes `os::macos::Sandboxfs` to specify a list of mounts
immediately on startup, rather than starting the "sandboxfs" thread with
no mounts and then reconfiguring at runtime via `mount()` and
`unmount()` methods. We also remove the code for generating unmap RPC
messages because dropping the `Sandboxfs` struct at the end of the
thread will automatically unmount all paths anyway.